context working correctly
