+++ homeworks/emiliano-acerbi/19-react-api/src/main.jsx
@@ -1,10 +1,13 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App'
-import './index.css'
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App';
+import CalculatorProvider from './context/CalculatorContext';
+import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>
-)
+    <CalculatorProvider>
+      <App />
+    </CalculatorProvider>
+  </React.StrictMode>,
+);
