--- homeworks/emiliano-acerbi/19-react-api/src/utils/numberClick.js
@@ -1,5 +1,0 @@
-export const numberClick = (btnValue, calcNum) => {
-  const inputToString = btnValue.toString();
-  const inputValue = Number(calcNum + inputToString);
-  return inputValue;
-};
